fix toolbar menu
