+++ README.md
@@ -21,7 +21,7 @@ This repository holds the source code examples for [Brightlayer UI design patter
 | [Contextual Action Bar](https://brightlayer-ui.github.io/patterns/appbar)                     | Contextual action bar to provide contextual actions to selected items.                                                                                                                                                | [CodeSandbox](https://codesandbox.io/s/github/brightlayer-ui/react-design-patterns/tree/master?file=/src/pages/app-bar/contextual-action/index.tsx&initialpath=/contextual-action)                                  |
 | [Contextual Spinners](https://brightlayer-ui.github.io/patterns/loading)                      | If you need to wait for a server response to validate inputs in a form (after clicking the submit button), you should disable the submit button and replace its label with a spinner.                                 | [CodeSandbox](https://codesandbox.io/s/github/brightlayer-ui/react-design-patterns/tree/master?file=/src/pages/loading-waiting-states/contextual-spinner/ContextualSpinner.tsx&initialpath=/contextual-spinner)     |
 | [Data List](https://brightlayer-ui.github.io/patterns/lists)                                  | A basic list dynamically built from a JSON data object.                                                                                                                                                               | [CodeSandbox](https://codesandbox.io/s/github/brightlayer-ui/react-design-patterns/tree/master?file=/src/pages/lists/data-list/index.tsx&initialpath=/data-list)                                                    |
-| [Dropdown Toolbar](https://brightlayer-ui.github.io/patterns/appbar)                          | Dropdown Toolbar transforms the simple text subtitle of the app bar into a dropdown menu that can be used to configure the page.                                                                                      | [CodeSandbox](https://codesandbox.io/s/github/brightlayer-ui/react-design-patterns/tree/master?file=/src/pages/app-bar/dropdown-toolbar/index.tsx&initialpath=/dropdown-toolbar)                                    |
+| [Toolbar Menu](https://brightlayer-ui.github.io/patterns/appbar)                          | The `ToolbarMenu` component is used to display a dropdown menu with label, primarily inside of `<Toolbar>` or Card Header elements.                                                                                      | [CodeSandbox](https://codesandbox.io/s/github/brightlayer-ui/react-design-patterns/tree/master?file=/src/pages/app-bar/toolbar-menu/index.tsx&initialpath=/toolbar-menu)                                    |
 | [Dynamic Stepper](https://brightlayer-ui.github.io/patterns/steppers)                         | A dynamic stepper that allows the user to create an arbitrary number of steps for a procedure.                                                                                                                        | [CodeSandbox](https://codesandbox.io/s/github/brightlayer-ui/react-design-patterns/tree/master?file=/src/pages/dynamic-stepper/index.tsx&initialpath=/dynamic-stepper)                                              |
 | [Fixed Length Passcode](https://brightlayer-ui.github.io/patterns/forms)                      | Form validation starts after specific length of passcode                                                                                                                                                              | [CodeSandbox](https://codesandbox.io/s/github/brightlayer-ui/react-design-patterns/tree/master?file=/src/pages/form-validation/fixed-length-passcode/index.tsx&initialpath=/fixed-length-passcode)                  |
 | [Forms In a List](https://brightlayer-ui.github.io/patterns/forms)                            | Input validations in a list                                                                                                                                                                                           | [CodeSandbox](https://codesandbox.io/s/github/brightlayer-ui/react-design-patterns/tree/master?file=/src/pages/form-validation/list/index.tsx&initialpath=/in-a-list)                                               |
