+++ src/pages/app-bar/toolbar-menu/toolbar-menu.test.tsx
@@ -1,22 +1,27 @@
 import React from 'react';
 import { cleanup, render } from '@testing-library/react';
-import { BluiDropdownToolbar } from '.';
+import { BluiToolbarMenu } from '.';
 import Enzyme from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import { Reducer } from '../../../redux/reducers';
+import { createTheme, ThemeProvider } from '@mui/material';
+import * as BLUIThemes from '@brightlayer-ui/react-themes';
+const theme = createTheme(BLUIThemes.blue);
 
 Enzyme.configure({ adapter: new Adapter() });
 const store = createStore(Reducer());
 
-describe('Blui dropdown toolbar', () => {
+describe('Blui toolbar menu', () => {
     afterEach(cleanup);
     it('renders without crashing', () => {
         render(
-            <Provider store={store}>
-                <BluiDropdownToolbar />
-            </Provider>
+            <ThemeProvider theme={theme}>
+                <Provider store={store}>
+                    <BluiToolbarMenu />
+                </Provider>
+            </ThemeProvider>
         );
     });
 });
